requirements.txt
# --- Scraping & ingestion ---
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
playwright==1.44.0          # headless browser for JS-rendered pages
tenacity==8.3.0             # retry logic
fake-useragent==1.5.1

# --- Data handling ---
pandas==2.2.2
pyarrow==16.1.0             # Parquet read/write

# --- dbt ---
dbt-core==1.8.3
dbt-duckdb==1.8.1           # local transformations
dbt-postgres==1.8.2         # write to Supabase

# --- DuckDB ---
duckdb==0.10.3

# --- Warehouse (Supabase / PostgreSQL) ---
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# --- Orchestration ---
apache-airflow==2.9.1

# --- Dashboard ---
streamlit==1.35.0
plotly==5.22.0
altair==5.3.0

# --- Utilities ---
python-dotenv==1.0.1
pydantic==2.7.4
loguru==0.7.2
tqdm==4.66.4
